created: delete permissions flow

+++ backend/src/resolvers/Mutation.js
@@ -2,7 +2,7 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
-const { makeANiceEmail, transport } = require('../mail');
+const { transport, makeANiceEmail } = require('../mail');
 const { hasPermission } = require('../utils');
 
 const Mutations = {
@@ -14,7 +14,7 @@ const Mutations = {
     const item = await ctx.db.mutation.createItem(
       {
         data: {
-          // This is how we create a relationship between the item and the User
+          // This is how to create a relationship between the Item and the User
           user: {
             connect: {
               id: ctx.request.userId,
@@ -31,7 +31,7 @@ const Mutations = {
   updateItem(parent, args, ctx, info) {
     // first take a copy of the updates
     const updates = { ...args };
-    // remove the ID from the update
+    // remove the ID from the updates
     delete updates.id;
     // run the update method
     return ctx.db.mutation.updateItem(
@@ -40,24 +40,34 @@ const Mutations = {
         where: {
           id: args.id,
         },
-      }, info,
+      },
+      info,
     );
   },
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
-    // 1.  find the item
-    const item = await ctx.db.query.item({ where }, '{id title}');
-    // 2.TODO: check to see if they own that item or have permisions
-
-    // delete the item
+    // 1. find the item
+    const item = await ctx.db.query.item({ where }, '{ id title user { id }}');
+    // 2. check to see if a user is logged in
+    const user = ctx.request.userId;
+    if (!user) {
+      throw new Error('You do not have permission to do this!!!');
+    }
+    // 3. Check if they own that item, or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
+    // 4. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
   async signup(parent, args, ctx, info) {
-    // Lowercase their email
+    // lowercase their email
     args.email = args.email.toLowerCase();
     // hash their password
     const password = await bcrypt.hash(args.password, 10);
-    // create the user in the db
+    // create the user in the database
     const user = await ctx.db.mutation.createUser(
       {
         data: {
@@ -65,16 +75,17 @@ const Mutations = {
           password,
           permissions: { set: ['USER'] },
         },
-      }, info,
+      },
+      info,
     );
-    // create JWT Token
+    // create the JWT token for them
     const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
-    // set the JWT as a cookie on the res
+    // We set the jwt as a cookie on the response
     ctx.response.cookie('token', token, {
       httpOnly: true,
       maxAge: 1000 * 60 * 60 * 24 * 365, // 1 year cookie
     });
-    // return the user to the browser
+    // Finalllllly we return the user to the browser
     return user;
   },
   async signin(parent, { email, password }, ctx, info) {
@@ -83,57 +94,59 @@ const Mutations = {
     if (!user) {
       throw new Error(`No such user found for email ${email}`);
     }
-    // 2. check id their password is correct
+    // 2. Check if their password is correct
     const valid = await bcrypt.compare(password, user.password);
     if (!valid) {
-      throw new Error('invalid password');
+      throw new Error('Invalid Password!');
     }
-    // 3. generate JWT Token
+    // 3. generate the JWT Token
     const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
-    // 4. set the cookie with the token
+    // 4. Set the cookie with the token
     ctx.response.cookie('token', token, {
       httpOnly: true,
-      maxAge: 1000 * 60 * 60 * 24 * 365, // 1 year cookie
+      maxAge: 1000 * 60 * 60 * 24 * 365,
     });
-    // 5. return the user.
+    // 5. Return the user
     return user;
   },
   signout(parent, args, ctx, info) {
     ctx.response.clearCookie('token');
-    return { message: 'GoodBye' };
+    return { message: 'Goodbye!' };
   },
   async requestReset(parent, args, ctx, info) {
-    // 1. Check is this is a real user
+    // 1. Check if this is a real user
     const user = await ctx.db.query.user({ where: { email: args.email } });
     if (!user) {
-      throw new Error(`No such User found for email ${args.email}`);
+      throw new Error(`No such user found for email ${args.email}`);
     }
     // 2. Set a reset token and expiry on that user
     const randomBytesPromiseified = promisify(randomBytes);
     const resetToken = (await randomBytesPromiseified(20)).toString('hex');
-    const resetTokenExpiry = Date.now() * 3600000;
+    const resetTokenExpiry = Date.now() + 3600000; // 1 hour from now
     const res = await ctx.db.mutation.updateUser({
       where: { email: args.email },
       data: { resetToken, resetTokenExpiry },
     });
     // 3. Email them that reset token
     const mailRes = await transport.sendMail({
-      from: 'jacob@cool.com',
+      from: 'wes@wesbos.com',
       to: user.email,
       subject: 'Your Password Reset Token',
-      html: makeANiceEmail(`Your Password reset token is here!
+      html: makeANiceEmail(`Your Password Reset Token is here!
       \n\n
-      <a href="${process.env.FRONTEND_URL}/reset?resetToken=${resetToken}">Click Here to Reset your Password</a>`),
+      <a href="${process.env
+    .FRONTEND_URL}/reset?resetToken=${resetToken}">Click Here to Reset</a>`),
     });
-    // Return the Message
-    return { message: 'thanks!' };
+
+    // 4. Return the message
+    return { message: 'Thanks!' };
   },
   async resetPassword(parent, args, ctx, info) {
-    // 1. Check if the passwords match
+    // 1. check if the passwords match
     if (args.password !== args.confirmPassword) {
-      throw new Error('Hey! your password does not match!');
+      throw new Error("Yo Passwords don't match!");
     }
-    // 2. Check if it is a legit reset token
+    // 2. check if its a legit reset token
     // 3. Check if its expired
     const [user] = await ctx.db.query.users({
       where: {
@@ -142,11 +155,11 @@ const Mutations = {
       },
     });
     if (!user) {
-      throw new Error('This token is either expired or invalid');
+      throw new Error('This token is either invalid or expired!');
     }
-    // 4. Hash theur new password
+    // 4. Hash their new password
     const password = await bcrypt.hash(args.password, 10);
-    // 5. Save the new password to the User and remove old resetToken
+    // 5. Save the new password to the user and remove old resetToken fields
     const updatedUser = await ctx.db.mutation.updateUser({
       where: { email: user.email },
       data: {
@@ -160,25 +173,28 @@ const Mutations = {
     // 7. Set the JWT cookie
     ctx.response.cookie('token', token, {
       httpOnly: true,
-      maxAge: 1000 * 60 * 60 * 24 * 365, // 1 year cookie
+      maxAge: 1000 * 60 * 60 * 24 * 365,
     });
-    // 8. return the new User
+    // 8. return the new user
     return updatedUser;
   },
   async updatePermissions(parent, args, ctx, info) {
-    // 1. check if they are logged in
+    // 1. Check if they are logged in
     if (!ctx.request.userId) {
-      throw new Error('you must be logged in');
+      throw new Error('You must be logged in!');
     }
     // 2. Query the current user
-    const currentUser = await ctx.db.query.user({
-      where: {
-        id: ctx.request.userId,
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId,
+        },
       },
-    }, info);
-    // 3. check if they have permissions to change
+      info,
+    );
+    // 3. Check if they have permissions to do this
     hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
-    // 4. update User Permissions
+    // 4. Update the permissions
     return ctx.db.mutation.updateUser(
       {
         data: {
@@ -189,7 +205,8 @@ const Mutations = {
         where: {
           id: args.userId,
         },
-      }, info,
+      },
+      info,
     );
   },
 };
